Changed dimensions int to intp.

+++ Lib/special/ufunc_extras.h
@@ -1,4 +1,4 @@
-#include "Python.h"
+#include "scipy/arrayobject.h"
 
 typedef int    IntFunc_d_dd(double x, double *y, double *z);
 typedef int    IntFunc_d_dddd(double v, double *w, double *x, double *y, double *z);
@@ -23,56 +23,56 @@ typedef double DoubleFunc_dddd_dd(double w, double x, double y, double z, double
 typedef double DoubleFunc_dddi_dd(double w, double x, double y, int z, double *err);
 
 
-extern void PyUFunc_f_ff_As_d_dd(char **args, int *dimensions, int *steps, void *func);
-extern void PyUFunc_d_dd(char **args, int *dimensions, int *steps, void *func);
-extern void PyUFunc_F_FF_As_D_DD(char **args, int *dimensions, int *steps, void *func);
-extern void PyUFunc_D_DD(char **args, int *dimensions, int *steps, void *func);
+extern void PyUFunc_f_ff_As_d_dd(char **args, intp *dimensions, int *steps, void *func);
+extern void PyUFunc_d_dd(char **args, intp *dimensions, int *steps, void *func);
+extern void PyUFunc_F_FF_As_D_DD(char **args, intp *dimensions, int *steps, void *func);
+extern void PyUFunc_D_DD(char **args, intp *dimensions, int *steps, void *func);
 
-extern void PyUFunc_f_FF_As_d_DD(char **args, int *dimensions, int *steps, void *func);
-extern void PyUFunc_d_DD(char **args, int *dimensions, int *steps, void *func);
+extern void PyUFunc_f_FF_As_d_DD(char **args, intp *dimensions, int *steps, void *func);
+extern void PyUFunc_d_DD(char **args, intp *dimensions, int *steps, void *func);
 
-extern void PyUFunc_ff_ff_As_dd_dd(char **args, int *dimensions, int *steps, void *func);
-extern void PyUFunc_dd_dd(char **args, int *dimensions, int *steps, void *func);
-extern void PyUFunc_f_ffff_As_d_dddd(char **args, int *dimensions, int *steps, void *func);
-extern void PyUFunc_d_dddd(char **args, int *dimensions, int *steps, void *func);
-extern void PyUFunc_f_ffff_As_D_DDDD(char **args, int *dimensions, int *steps, void *func);
-extern void PyUFunc_d_dddd_As_D_DDDD(char **args, int *dimensions, int *steps, void *func);
-extern void PyUFunc_F_FFFF_As_D_DDDD(char **args, int *dimensions, int *steps, void *func);
-extern void PyUFunc_D_DDDD(char **args, int *dimensions, int *steps, void *func);
+extern void PyUFunc_ff_ff_As_dd_dd(char **args, intp *dimensions, int *steps, void *func);
+extern void PyUFunc_dd_dd(char **args, intp *dimensions, int *steps, void *func);
+extern void PyUFunc_f_ffff_As_d_dddd(char **args, intp *dimensions, int *steps, void *func);
+extern void PyUFunc_d_dddd(char **args, intp *dimensions, int *steps, void *func);
+extern void PyUFunc_f_ffff_As_D_DDDD(char **args, intp *dimensions, int *steps, void *func);
+extern void PyUFunc_d_dddd_As_D_DDDD(char **args, intp *dimensions, int *steps, void *func);
+extern void PyUFunc_F_FFFF_As_D_DDDD(char **args, intp *dimensions, int *steps, void *func);
+extern void PyUFunc_D_DDDD(char **args, intp *dimensions, int *steps, void *func);
 
-extern void PyUFunc_f_FFFF_As_d_DDDD(char **args, int *dimensions, int *steps, void *func);
-extern void PyUFunc_d_DDDD(char **args, int *dimensions, int *steps, void *func);
+extern void PyUFunc_f_FFFF_As_d_DDDD(char **args, intp *dimensions, int *steps, void *func);
+extern void PyUFunc_d_DDDD(char **args, intp *dimensions, int *steps, void *func);
 
 
-extern void PyUFunc_ff_ffff_As_dd_dddd(char **args, int *dimensions, int *steps, void *func);
-extern void PyUFunc_dd_dddd(char **args, int *dimensions, int *steps, void *func);
-extern void PyUFunc_fff_f_As_ddd_d(char **args, int *dimensions, int *steps, void *func);
-extern void PyUFunc_ddd_d(char **args, int *dimensions, int *steps, void *func);
-extern void PyUFunc_fff_ff_As_ddd_dd(char **args, int *dimensions, int *steps, void *func);
-extern void PyUFunc_ddd_dd(char **args, int *dimensions, int *steps, void *func);
+extern void PyUFunc_ff_ffff_As_dd_dddd(char **args, intp *dimensions, int *steps, void *func);
+extern void PyUFunc_dd_dddd(char **args, intp *dimensions, int *steps, void *func);
+extern void PyUFunc_fff_f_As_ddd_d(char **args, intp *dimensions, int *steps, void *func);
+extern void PyUFunc_ddd_d(char **args, intp *dimensions, int *steps, void *func);
+extern void PyUFunc_fff_ff_As_ddd_dd(char **args, intp *dimensions, int *steps, void *func);
+extern void PyUFunc_ddd_dd(char **args, intp *dimensions, int *steps, void *func);
 
-extern void PyUFunc_ff_f_As_id_d(char **args, int *dimensions, int *steps, void *func);
-extern void PyUFunc_dd_d_As_id_d(char **args, int *dimensions, int *steps, void *func);
+extern void PyUFunc_ff_f_As_id_d(char **args, intp *dimensions, int *steps, void *func);
+extern void PyUFunc_dd_d_As_id_d(char **args, intp *dimensions, int *steps, void *func);
 
-extern void PyUFunc_ff_f_As_dD_D(char **args, int *dimensions, int *steps, void *func);
-extern void PyUFunc_dd_d_As_dD_D(char **args, int *dimensions, int *steps, void *func);
-extern void PyUFunc_fF_F_As_dD_D(char **args, int *dimensions, int *steps, void *func);
-extern void PyUFunc_dD_D(char **args, int *dimensions, int *steps, void *func);
+extern void PyUFunc_ff_f_As_dD_D(char **args, intp *dimensions, int *steps, void *func);
+extern void PyUFunc_dd_d_As_dD_D(char **args, intp *dimensions, int *steps, void *func);
+extern void PyUFunc_fF_F_As_dD_D(char **args, intp *dimensions, int *steps, void *func);
+extern void PyUFunc_dD_D(char **args, intp *dimensions, int *steps, void *func);
 
-extern void PyUFunc_ffF_F_As_ddD_D(char **args, int *dimensions, int *steps, void *func);
-extern void PyUFunc_ddD_D(char **args, int *dimensions, int *steps, void *func);
-extern void PyUFunc_fffF_F_As_dddD_D(char **args, int *dimensions, int *steps, void *func);
-extern void PyUFunc_dddD_D(char **args, int *dimensions, int *steps, void *func);
+extern void PyUFunc_ffF_F_As_ddD_D(char **args, intp *dimensions, int *steps, void *func);
+extern void PyUFunc_ddD_D(char **args, intp *dimensions, int *steps, void *func);
+extern void PyUFunc_fffF_F_As_dddD_D(char **args, intp *dimensions, int *steps, void *func);
+extern void PyUFunc_dddD_D(char **args, intp *dimensions, int *steps, void *func);
 
-extern void PyUFunc_fff_f_As_iid_d(char **args, int *dimensions, int *steps, void *func);
-extern void PyUFunc_ddd_d_As_iid_d(char **args, int *dimensions, int *steps, void *func);
-extern void PyUFunc_ffff_f_As_dddd_d(char **args, int *dimensions, int *steps, void *func);
-extern void PyUFunc_dddd_d(char **args, int *dimensions, int *steps, void *func);
-extern void PyUFunc_ffff_ff_As_dddd_dd(char **args, int *dimensions, int *steps, void *func);
-extern void PyUFunc_dddd_dd(char **args, int *dimensions, int *steps, void *func);
-extern void PyUFunc_fffff_ff_As_ddddd_dd(char **args, int *dimensions, int *steps, void *func);
-extern void PyUFunc_ddddd_dd(char **args, int *dimensions, int *steps, void *func);
-extern void PyUFunc_ffff_ff_As_dddi_dd(char **args, int *dimensions, int *steps, void *func);
-extern void PyUFunc_dddd_dd_As_dddi_dd(char **args, int *dimensions, int *steps, void *func);
+extern void PyUFunc_fff_f_As_iid_d(char **args, intp *dimensions, int *steps, void *func);
+extern void PyUFunc_ddd_d_As_iid_d(char **args, intp *dimensions, int *steps, void *func);
+extern void PyUFunc_ffff_f_As_dddd_d(char **args, intp *dimensions, int *steps, void *func);
+extern void PyUFunc_dddd_d(char **args, intp *dimensions, int *steps, void *func);
+extern void PyUFunc_ffff_ff_As_dddd_dd(char **args, intp *dimensions, int *steps, void *func);
+extern void PyUFunc_dddd_dd(char **args, intp *dimensions, int *steps, void *func);
+extern void PyUFunc_fffff_ff_As_ddddd_dd(char **args, intp *dimensions, int *steps, void *func);
+extern void PyUFunc_ddddd_dd(char **args, intp *dimensions, int *steps, void *func);
+extern void PyUFunc_ffff_ff_As_dddi_dd(char **args, intp *dimensions, int *steps, void *func);
+extern void PyUFunc_dddd_dd_As_dddi_dd(char **args, intp *dimensions, int *steps, void *func);
